+++ crates/rhdl-fpga/examples/stream_map.rs
@@ -1,6 +1,7 @@
 use std::iter::repeat_n;
 
 use rhdl::prelude::*;
+use rhdl_fpga::stream::testing::sink_from_fn::SinkView;
 use rhdl_fpga::{
     core::slice::lsbs,
     rng::xorshift::XorShift128,
@@ -19,8 +20,8 @@ fn main() -> Result<(), RHDLError> {
     let a_rng = XorShift128::default().map(|x| b4((x & 0xF) as u128));
     let mut b_rng = a_rng.clone();
     let a_rng = stalling(a_rng, 0.23);
-    let consume = move |data: Option<b2>| {
-        if let Some(data) = data {
+    let consume = move |v: SinkView<b2>| {
+        if let Some(data) = v.accepted {
             let orig = b_rng.next().unwrap();
             let orig_lsb = lsbs::<2, 4>(orig);
             assert_eq!(data, orig_lsb);