+++ crates/rhdl-fpga/examples/tee.rs
@@ -1,6 +1,7 @@
 use std::iter::repeat_n;
 
 use rhdl::prelude::*;
+use rhdl_fpga::stream::testing::sink_from_fn::SinkView;
 use rhdl_fpga::{
     rng::xorshift::XorShift128,
     stream::{
@@ -45,15 +46,15 @@ fn main() -> Result<(), RHDLError> {
     let mut c_rng = a_rng.clone();
     let mut d_rng = a_rng.clone();
     let a_rng = stalling(a_rng, 0.23);
-    let consume_s = move |data| {
-        if let Some(data) = data {
+    let consume_s = move |v: SinkView<_>| {
+        if let Some(data) = v.accepted {
             let validation = c_rng.next().unwrap();
             assert_eq!(data, validation.0);
         }
         rand::random::<f64>() > 0.2
     };
-    let consume_t = move |data| {
-        if let Some(data) = data {
+    let consume_t = move |v: SinkView<_>| {
+        if let Some(data) = v.accepted {
             let validation = d_rng.next().unwrap();
             assert_eq!(data, validation.1);
         }