+++ crates/rhdl-fpga/src/axi4lite/register/testing/single_streaming_writes.rs
@@ -8,7 +8,11 @@ use crate::{
     },
     core::dff::DFF,
     rng::xorshift::{XorShift, XorShift128},
-    stream::testing::{sink_from_fn::SinkFromFn, source_from_fn::SourceFromFn, utils::stalling},
+    stream::testing::{
+        sink_from_fn::{SinkFromFn, SinkView},
+        source_from_fn::SourceFromFn,
+        utils::stalling,
+    },
 };
 
 #[derive(Clone, Synchronous, SynchronousDQ)]
@@ -35,8 +39,8 @@ impl Default for Fixture {
         });
         let cmd = stalling(cmd, 0.23);
         // For the write sink, we expect all writes to succeed
-        let acceptor = |x: Option<WriteResult>| {
-            if let Some(res) = x {
+        let acceptor = |v: SinkView<WriteResult>| {
+            if let Some(res) = v.accepted {
                 assert_eq!(res, Ok(()));
             }
             rand::random_bool(0.85)