+++ crates/rhdl-fpga/src/stream/map.rs
@@ -152,6 +152,7 @@ mod tests {
     };
 
     use super::*;
+    use crate::stream::testing::sink_from_fn::SinkView;
 
     #[kernel]
     fn map_item(_cr: ClockReset, t: b4) -> b2 {
@@ -230,8 +231,8 @@ mod tests {
         let a_rng = XorShift128::default().map(|x| b4((x & 0xF) as u128));
         let mut b_rng = a_rng.clone();
         let a_rng = stalling(a_rng, 0.23);
-        let consume = move |data: Option<b2>| {
-            if let Some(data) = data {
+        let consume = move |v: SinkView<b2>| {
+            if let Some(data) = v.accepted {
                 let orig = b_rng.next().unwrap();
                 let orig_lsb = lsbs::<2, 4>(orig);
                 assert_eq!(data, orig_lsb);